pyproject.toml
[project]
name = "daily-paper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.0",
    "bs4>=0.0.2",
    "feedparser>=6.0.12",
    "pyyaml>=6.0.0",
    "requests>=2.32.5",
]
